增加音乐播放功能

+++ Ribbon1.cs
@@ -5,39 +5,52 @@ using System.Diagnostics;
 using System.IO;
 using System.Windows.Forms;
 using Excel = Microsoft.Office.Interop.Excel;
+using NAudio.Wave;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.Office.Interop.Excel;
+using NAudio.CoreAudioApi;
 
 namespace ExcelAddIn
 {
     public partial class Ribbon1
     {
+        
+        //是否执行读文件名功能标识
         private int readFile;
 
         private void Ribbon1_Load(object sender, RibbonUIEventArgs e)
         {
-            select_f_or_d.Checked = false;
-            select_f_or_d.Label = "改文件名";
-            select_f_or_d.ShowLabel = false;
+            Select_f_or_d.Checked = false;
+            Select_f_or_d.Label = "改文件名";
+            Select_f_or_d.ShowLabel = false;
             switch_FD_label.Label = "文件名";
             readFile = 0;
+            playbackMode = PlaybackMode.Sequential;
+            Mode_btuuon.Label = "顺序播放";
+            Mode_btuuon.Image = Properties.Resources.order_play;
+            currentPlayState = PlaybackState.Stopped;
+            Select_mp3_button.Image = Properties.Resources.no_open_fold;
+            musicFiles.Clear();
         }
 
 
         //表操作按钮
-        private void excel_extend_Click(object sender, RibbonControlEventArgs e)
+        private void Excel_extend_Click(object sender, RibbonControlEventArgs e)
         {
             Form1 form1 = new Form1();
             form1.ShowDialog();
         }
 
         //邮件群发按钮
-        private void send_mail_Click(object sender, RibbonControlEventArgs e)
+        private void Send_mail_Click(object sender, RibbonControlEventArgs e)
         {
             Form2 form2 = new Form2();
             form2.ShowDialog();
         }
 
         //指定字段名所处的列
-        private int getUsedRangeColumn(string targetColumn)
+        private int GetUsedRangeColumn(string targetColumn)
         {
             for (int n = 1; n <= ThisAddIn.app.ActiveSheet.UsedRange.Columns.Count; n++)
             {
@@ -52,7 +65,7 @@ namespace ExcelAddIn
         string get_directory_path;
 
         //批读文件名
-        private void files_read_Click(object sender, RibbonControlEventArgs e)
+        private void Files_read_Click(object sender, RibbonControlEventArgs e)
         {
             Excel.Workbook workbook = ThisAddIn.app.ActiveWorkbook;
             ThisAddIn.app.DisplayAlerts = false;
@@ -86,7 +99,7 @@ namespace ExcelAddIn
                 Excel.Worksheet worksheet = workbook.Worksheets.Add();
                 worksheet.Name = "_rename";
                 worksheet.Activate();
-                switch (select_f_or_d.Checked)
+                switch (Select_f_or_d.Checked)
                 {
                     case false:
                         worksheet.Cells[1, 1] = "路径";
@@ -139,7 +152,7 @@ namespace ExcelAddIn
         }
 
         //批量重命名
-        private void file_rename_Click(object sender, RibbonControlEventArgs e)
+        private void File_rename_Click(object sender, RibbonControlEventArgs e)
         {
             Excel.Workbook workbook = ThisAddIn.app.ActiveWorkbook;
             ThisAddIn.app.ScreenUpdating = false;
@@ -153,35 +166,43 @@ namespace ExcelAddIn
                     string cell1 = workbook.Worksheets["_rename"].Cells[i, 1].Value;
                     string cell2 = workbook.Worksheets["_rename"].Cells[i, 2].Value;
                     string cell3 = workbook.Worksheets["_rename"].Cells[i, 3].Value;
-                    string full_path = cell1;
-                    string old_name = Path.Combine(cell1, cell2);
-                    string new_name = Path.Combine(cell1, cell3);
-                    switch (select_f_or_d.Checked == true)
+                    if (string.IsNullOrEmpty(cell1) || string.IsNullOrEmpty(cell2) || string.IsNullOrEmpty(cell3))
                     {
-                        case false:
-                            int exist_file = 0;
-                            if (old_name != new_name)
-                            {
-                                while (File.Exists(new_name))
+                        MessageBox.Show($"第{i}行有空格，请检查！该行对应文件将不被改名");
+                        continue;
+                    }
+                    else
+                    {
+                        string full_path = cell1;
+                        string old_name = Path.Combine(cell1, cell2);
+                        string new_name = Path.Combine(cell1, cell3);
+                        switch (Select_f_or_d.Checked == true)
+                        {
+                            case false:
+                                int exist_file = 0;
+                                if (old_name != new_name)
                                 {
-                                    exist_file++;
-                                    new_name = Path.Combine(cell1, Path.GetFileNameWithoutExtension(new_name) + "(" + exist_file.ToString() + ")" + Path.GetExtension(new_name));
+                                    while (File.Exists(new_name))
+                                    {
+                                        exist_file++;
+                                        new_name = Path.Combine(cell1, Path.GetFileNameWithoutExtension(new_name) + "(" + exist_file.ToString() + ")" + Path.GetExtension(new_name));
+                                    }
+                                    File.Move(old_name, new_name);
                                 }
-                                File.Move(old_name, new_name);
-                            }
-                            break;
-                        case true:
-                            int exist_fold = 0;
-                            if (old_name != new_name)
-                            {
-                                while (Directory.Exists(new_name))
+                                break;
+                            case true:
+                                int exist_fold = 0;
+                                if (old_name != new_name)
                                 {
-                                    exist_fold++;
-                                    new_name = Path.Combine(cell1, cell3 + "(" + exist_fold.ToString() + ")");
+                                    while (Directory.Exists(new_name))
+                                    {
+                                        exist_fold++;
+                                        new_name = Path.Combine(cell1, cell3 + "(" + exist_fold.ToString() + ")");
+                                    }
+                                    Directory.Move(old_name, new_name);
                                 }
-                                Directory.Move(old_name, new_name);
-                            }
-                            break;
+                                break;
+                        }
                     }
                 }
 
@@ -206,25 +227,25 @@ namespace ExcelAddIn
         }
 
         //文件目录选项
-        private void select_f_or_d_Click(object sender, RibbonControlEventArgs e)
+        private void Select_f_or_d_Click(object sender, RibbonControlEventArgs e)
         {
-            if (select_f_or_d.Checked == true)
+            if (Select_f_or_d.Checked == true)
             {
-                select_f_or_d.Image = ExcelAddIn.Properties.Resources.Radio_Button_on;
-                select_f_or_d.Label = "改文件夹名";
-                select_f_or_d.ShowLabel = false;
+                Select_f_or_d.Image = ExcelAddIn.Properties.Resources.Radio_Button_on;
+                Select_f_or_d.Label = "改文件夹名";
+                Select_f_or_d.ShowLabel = false;
                 switch_FD_label.Label = "目录名";
             }
             else
             {
-                select_f_or_d.Image = ExcelAddIn.Properties.Resources.Radio_Button_off;
-                select_f_or_d.Label = "改文件名";
-                select_f_or_d.ShowLabel = false;
+                Select_f_or_d.Image = ExcelAddIn.Properties.Resources.Radio_Button_off;
+                Select_f_or_d.Label = "改文件名";
+                Select_f_or_d.ShowLabel = false;
                 switch_FD_label.Label = "文件名";
             }
         }
 
-        private void rename_mp3_Click(object sender, RibbonControlEventArgs e)
+        private void Rename_mp3_Click(object sender, RibbonControlEventArgs e)
         {
             Form3 form3 = new Form3();
             form3.ShowDialog();
@@ -242,5 +263,325 @@ namespace ExcelAddIn
             }
             return false;
         }
+
+
+        //音乐播放模式
+        private enum PlaybackMode
+        {
+            Sequential,
+            SingleLoop,
+            AllLoop
+        }
+        //音乐播放状态
+        private enum PlaybackState
+        {
+            Stopped,
+            Playing,
+            Paused
+        }
+        //初始化音乐播放模式
+        private PlaybackMode playbackMode=PlaybackMode.Sequential;
+        //初始化音乐播放状态
+        private PlaybackState currentPlayState=PlaybackState.Stopped;
+
+        //实例化WaveOutEvent对象
+        private WaveOutEvent waveOutEvent;        
+        //实例化AudioFileReader对象
+        private AudioFileReader audioFile=null;
+        //音乐播放列表
+        private List<string> musicFiles = new List<string>();
+        //当前播放歌曲序号
+        private int currentSongIndex = -1;
+
+        //选择音乐文件夹
+        private void Select_mp3_button_Click(object sender, RibbonControlEventArgs e)
+        {
+            musicFiles.Clear();
+            DisposeWavePlayer();
+            folderBrowserDialog1.Description = "请选择音乐文件所在文件夹";
+            if (folderBrowserDialog1.ShowDialog() == DialogResult.OK)
+            {
+                LoadMusicFiles(folderBrowserDialog1.SelectedPath);                
+                if(musicFiles.Count == 0)
+                {
+                    Select_mp3_button.Image = Properties.Resources.no_open_fold;
+                    ThisAddIn.app.Application.StatusBar = "选择的文件夹中没有支持格式的音乐";
+                }
+                else
+                {
+                    currentSongIndex = 0;
+                    Select_mp3_button.Image = Properties.Resources.fold;
+                    ThisAddIn.app.Application.StatusBar = $"当前未播放音乐，共{musicFiles.Count}首音乐，第1首：{Path.GetFileName(musicFiles[currentSongIndex])}";
+                }
+            }
+            else
+            {
+                return;
+            }
+        }
+
+        //写入播放列表方法
+        private void LoadMusicFiles(string folderPath)
+        {
+            musicFiles.AddRange(Directory.GetFiles(folderPath, "*.mp3", SearchOption.AllDirectories));
+            musicFiles.AddRange(Directory.GetFiles(folderPath, "*.wav", SearchOption.AllDirectories));
+            musicFiles.AddRange(Directory.GetFiles(folderPath, "*.flac", SearchOption.AllDirectories));
+            musicFiles.AddRange(Directory.GetFiles(folderPath, "*.aiff", SearchOption.AllDirectories));
+            musicFiles.AddRange(Directory.GetFiles(folderPath, "*.wma", SearchOption.AllDirectories));
+            musicFiles.AddRange(Directory.GetFiles(folderPath, "*.aac", SearchOption.AllDirectories));
+            musicFiles.AddRange(Directory.GetFiles(folderPath, "*.g711", SearchOption.AllDirectories));
+            musicFiles.AddRange(Directory.GetFiles(folderPath, "*.mp4", SearchOption.AllDirectories));
+        }
+
+        //播放模式选择按钮
+        private void Mode_btuuon_Click(object sender, RibbonControlEventArgs e)
+        {
+            switch(playbackMode)
+            {
+                case PlaybackMode.Sequential:
+                    playbackMode = PlaybackMode.AllLoop;
+                    Mode_btuuon.Label = "全部循环";
+                    Mode_btuuon.ScreenTip = "全部循环";
+                    Mode_btuuon.Image = Properties.Resources.all_cycle;
+                    break;
+                case PlaybackMode.AllLoop:
+                    playbackMode=PlaybackMode.SingleLoop;
+                    Mode_btuuon.Label = "单曲循环";
+                    Mode_btuuon.ScreenTip = "单曲循环";
+                    Mode_btuuon.Image = Properties.Resources.single_cycle;
+                    break;
+                case PlaybackMode.SingleLoop:
+                    playbackMode = PlaybackMode.Sequential;
+                    Mode_btuuon.Label = "顺序播放";
+                    Mode_btuuon.ScreenTip = "顺序播放";
+                    Mode_btuuon.Image = Properties.Resources.order_play;
+                    break;
+            }
+        }
+
+        private bool isMusicEnded=false;
+
+        //播放按钮单击事件
+        private async void Play_button_Click(object sender, RibbonControlEventArgs e)
+        {
+            //timer1.Interval = 200;
+
+
+            if (musicFiles.Count == 0)
+                return;
+
+            syncContext = new SynchronizationContext();
+
+            isAutoContinue = true;
+
+            if (currentPlayState == PlaybackState.Stopped)
+            {
+                //timer1.Enabled = true;
+                await PlayMusic();
+            }
+            else if (currentPlayState == PlaybackState.Paused)
+            {
+                waveOutEvent.Play();
+                //timer1.Enabled = true;
+                currentPlayState = PlaybackState.Playing;
+                UpdateTrackInfo();
+                Play_button.Image = Properties.Resources.pause;
+                Play_button.ScreenTip = "暂停";
+            }
+            else
+            {
+                waveOutEvent?.Pause();
+                //timer1.Enabled = false;
+                currentPlayState = PlaybackState.Paused;
+                ThisAddIn.app.Application.StatusBar = "播放暂停中......";
+                Play_button.Image = Properties.Resources.play;
+                Play_button.ScreenTip = "播放";
+            }
+        }
+
+        //播放/暂停（方法）
+        private async Task PlayMusic()
+        {
+            
+            isMusicEnded = false;
+            if (currentPlayState == PlaybackState.Stopped)
+            {
+                currentPlayState = PlaybackState.Playing;
+                Play_button.Image = Properties.Resources.pause;
+                Play_button.ScreenTip = "暂停";
+                DisposeWavePlayer();
+                if (waveOutEvent == null)
+                {
+                    waveOutEvent = new WaveOutEvent();
+                    waveOutEvent.PlaybackStopped += OnPlaybackStopped;
+                }
+                if (audioFile == null)
+                {
+                    audioFile = new AudioFileReader(musicFiles[currentSongIndex]);
+                    waveOutEvent?.Init(audioFile);
+                }
+                UpdateTrackInfo();
+                waveOutEvent.Play();
+
+                await Task.Delay(200);
+
+                // 等待播放完成
+                while (!isMusicEnded)
+                {
+                    await Task.Delay(100);
+                }
+            }
+        }
+
+        private SynchronizationContext syncContext = SynchronizationContext.Current;
+
+        //播放完毕事件触发
+        private async void OnPlaybackStopped(object sender, StoppedEventArgs args)
+        {
+            DisposeWavePlayer();
+            isMusicEnded = true;
+            currentPlayState = PlaybackState.Stopped;
+
+            if (isAutoContinue == false)
+            {
+                ThisAddIn.app.Application.StatusBar = false;
+                currentSongIndex = 0;
+                return;
+            }
+            else
+            {
+                //// 在UI线程上更新UI
+                //syncContext.Send(state =>
+                //{
+                //    UpdateTrackInfo();
+                //}, null);
+
+                // 播放下一首歌曲
+                currentSongIndex++;
+                switch (playbackMode)
+                {
+
+                    case PlaybackMode.Sequential:
+                        if (currentSongIndex < musicFiles.Count)
+                        {
+                            await Task.Delay(1000);
+                            await PlayMusic();
+                        }
+                        else
+                        {
+                            currentSongIndex = 0;
+                            StopMusic();
+                            
+                        }                       
+                        break;
+
+                    case PlaybackMode.AllLoop:
+                        if (currentSongIndex == musicFiles.Count)
+                        {
+                            currentSongIndex = 0;
+                            
+                        }
+                        await Task.Delay(1000);
+                        await PlayMusic();
+                        break;
+                    case PlaybackMode.SingleLoop:
+                        currentSongIndex--;
+                        await Task.Delay(1000);
+                        await PlayMusic();
+                        break;
+                }
+            }
+        }
+
+        private bool isAutoContinue = true;
+
+        //停止按钮单击事件
+        private void Stop_button_Click(object sender, RibbonControlEventArgs e)
+        {
+            StopMusic();
+        }
+
+
+        private void StopMusic()
+        {
+            if (waveOutEvent != null)
+            {
+                waveOutEvent.Stop();
+                waveOutEvent.PlaybackStopped += OnPlaybackStopped;
+
+            }
+            if (musicFiles.Count > 0)
+            {
+                ThisAddIn.app.Application.StatusBar = $"播放已停止，共{musicFiles.Count}首音乐，可从第1首重新播放";
+            }
+            else
+            {
+                ThisAddIn.app.Application.StatusBar = false;
+            }
+            isAutoContinue = false;
+            //timer1.Enabled = false;
+            Play_button.Image = Properties.Resources.play;
+            Play_button.Label = "播放";
+        }
+
+        private async void Next_button_Click(object sender, RibbonControlEventArgs e)
+        {
+            if (musicFiles.Count != 0)
+            {
+                waveOutEvent.Stop();
+                currentSongIndex = (currentSongIndex + 1) % musicFiles.Count;
+                await PlayMusic();
+            }
+
+        }
+
+        private async void Previous_button_Click(object sender, RibbonControlEventArgs e)
+        {
+            if (musicFiles.Count != 0)
+            {
+                waveOutEvent.Stop();
+                currentSongIndex = (currentSongIndex - 1 + musicFiles.Count) % musicFiles.Count;
+                await PlayMusic();
+            }
+        }
+
+        //清理wavePlayer
+        private void DisposeWavePlayer()
+        {
+            waveOutEvent?.Dispose();
+            waveOutEvent = null;
+            audioFile?.Dispose();
+            audioFile = null;
+        }
+
+        //当前播放歌曲信息
+        private async  void UpdateTrackInfo()
+        {
+            if (audioFile != null && currentSongIndex < musicFiles.Count)
+            {
+                string trackInfo = await Task.Run(() =>
+                {
+                    return $"正在播放第{currentSongIndex + 1}首：{Path.GetFileName(musicFiles[currentSongIndex])}，" +
+                        $"时长：{audioFile.TotalTime.ToString(@"mm\:ss")}";
+                    //return $"正在播放第{currentSongIndex + 1}首：{Path.GetFileName(musicFiles[currentSongIndex])}，" +
+                    //    $"已播放时长：{audioFile.CurrentTime.ToString(@"mm\:ss")}，" + $"歌曲时长：{audioFile.TotalTime.ToString(@"mm\:ss")}";
+                });
+
+                ThisAddIn.app.Application.StatusBar = trackInfo;
+
+            }
+            else
+            {
+
+               ThisAddIn.app.Application.StatusBar = false;
+
+            }
+        }
+        
+        private void timer1_Tick(object sender, EventArgs e)
+        {
+           UpdateTrackInfo();            
+        }
     }
 }
+
